feat: replace `errgroup` with `ants` goroutine pool
This commit is an attempt for issue #7.

There seems to be a small performance improvement (less than 1s).

Before:

  time ./find-duplicates -o /dev/null /home

  real	0m23.034s
  user	2m27.127s
  sys	1m23.543s

After:

  time ./find-duplicates -o /dev/null /home

  real	0m22.236s
  user	2m27.485s
  sys	1m25.284s
