+++ go.mod
@@ -3,9 +3,9 @@ module github.com/twpayne/find-duplicates
 go 1.21
 
 require (
+	github.com/panjf2000/ants/v2 v2.9.0
 	github.com/spf13/pflag v1.0.5
 	github.com/zeebo/xxh3 v1.0.2
-	golang.org/x/sync v0.3.0
 )
 
 require github.com/klauspost/cpuid/v2 v2.0.9 // indirect
